Cognito/MultipleCognitoPOC/lambdaAuthorizer/requirements.txt
cffi==1.14.5
cryptography==3.4.5
pycparser==2.20
PyJWT==2.0.1
